start to abstract out the local client simulators so there can be many flavors
