+++ src/simularium/ClientSimulatorConnection.ts
@@ -1,213 +1,23 @@
 import jsLogger from "js-logger";
 import { ILogger } from "js-logger";
 
-import VisTypes from "./VisTypes";
-import {
-    VisDataMessage,
-    TrajectoryFileInfoV2,
-    EncodedTypeMapping,
-} from "./types";
+import { VisDataMessage, TrajectoryFileInfoV2 } from "./types";
 import { NetConnection } from "./NetConnection";
+import {
+    ClientMessageEnum,
+    ClientPlayBackType,
+} from "./localSimulators/IClientSimulator";
+import { createSimulator } from "./localSimulators/ClientSimulatorFactory";
 
-class CurveSim {
-    nCurves: number;
-    curveData: number[];
-    nPointsPerCurve: number;
-    currentFrame: number;
-    nTypes: number;
+const localSimulator = createSimulator({
+    name: "my curve sim",
+    type: "CURVESIM",
+    nCurves: 1000,
+    nTypes: 4,
+});
 
-    constructor(nCurves: number, nTypes: number) {
-        this.nCurves = nCurves;
-        this.nTypes = nTypes;
-        this.nPointsPerCurve = 5;
-        this.curveData = this.makeCurveBundle(nCurves, this.nPointsPerCurve);
-        this.currentFrame = 0;
-    }
-
-    private randomFloat(min, max) {
-        if (max === undefined) {
-            max = min;
-            min = 0;
-        }
-        return Math.random() * (max - min) + min;
-    }
-
-    private randomSpherePoint(x0, y0, z0, radius): number[] {
-        const u = Math.random();
-        const v = Math.random();
-        const theta = 2 * Math.PI * u;
-        const phi = Math.acos(2 * v - 1);
-        const x = x0 + radius * Math.sin(phi) * Math.cos(theta);
-        const y = y0 + radius * Math.sin(phi) * Math.sin(theta);
-        const z = z0 + radius * Math.cos(phi);
-        return [x, y, z];
-    }
-    private randomPtInBox(xmin, xmax, ymin, ymax, zmin, zmax) {
-        return [
-            this.randomFloat(xmin, xmax),
-            this.randomFloat(ymin, ymax),
-            this.randomFloat(zmin, zmax),
-        ];
-    }
-
-    private makeCurveBundle(nCurves, nPts) {
-        const curves: number[] = [];
-        let p: number[];
-        if (nPts === 3) {
-            for (let i = 0; i < nCurves; ++i) {
-                p = this.randomSpherePoint(0, 0, 0, 4.0);
-                curves.push(p[0]);
-                curves.push(p[1]);
-                curves.push(p[2]);
-                p = this.randomSpherePoint(0, 0, 0, 0.25);
-                curves.push(p[0]);
-                curves.push(p[1]);
-                curves.push(p[2]);
-                p = this.randomSpherePoint(0, 0, 0, 2.0);
-                curves.push(p[0]);
-                curves.push(p[1]);
-                curves.push(p[2]);
-            }
-        } else if (nPts === 5) {
-            for (let i = 0; i < nCurves; ++i) {
-                p = this.randomPtInBox(-4, -3, -2, 2, -2, 2);
-                curves.push(p[0]);
-                curves.push(p[1]);
-                curves.push(p[2]);
-                p = this.randomPtInBox(-2.5, -2, -1, 1, -1, 1);
-                curves.push(p[0]);
-                curves.push(p[1]);
-                curves.push(p[2]);
-                p = this.randomPtInBox(-1, 1, -0.5, 0.5, -0.5, 0.5);
-                curves.push(p[0]);
-                curves.push(p[1]);
-                curves.push(p[2]);
-                p = this.randomPtInBox(2, 2.5, -1, 1, -1, 1);
-                curves.push(p[0]);
-                curves.push(p[1]);
-                curves.push(p[2]);
-                p = this.randomPtInBox(3, 4, -2, 2, -2, 2);
-                curves.push(p[0]);
-                curves.push(p[1]);
-                curves.push(p[2]);
-            }
-        }
-        return curves;
-    }
-
-    public update(dt: number): VisDataMessage {
-        const nFloatsPerCurve = this.nPointsPerCurve * 3;
-        //const dt_adjusted = dt / 1000;
-        const amplitude = 0.05;
-        for (let ii = 0; ii < this.nCurves; ++ii) {
-            for (let jj = 0; jj < this.nPointsPerCurve; ++jj) {
-                this.curveData[
-                    ii * nFloatsPerCurve + jj * 3 + 0
-                ] += this.randomFloat(-amplitude, amplitude);
-                this.curveData[
-                    ii * nFloatsPerCurve + jj * 3 + 1
-                ] += this.randomFloat(-amplitude, amplitude);
-                this.curveData[
-                    ii * nFloatsPerCurve + jj * 3 + 2
-                ] += this.randomFloat(-amplitude, amplitude);
-            }
-        }
-        // fill agent data.
-        const agentData: number[] = [];
-        for (let ii = 0; ii < this.nCurves; ++ii) {
-            agentData.push(VisTypes.ID_VIS_TYPE_FIBER); // vis type
-            agentData.push(ii); // instance id
-            agentData.push(ii % this.nTypes); // type
-            agentData.push(0); // x
-            agentData.push(0); // y
-            agentData.push(0); // z
-            agentData.push(0); // rx
-            agentData.push(0); // ry
-            agentData.push(0); // rz
-            agentData.push(0.1); // collision radius
-            agentData.push(nFloatsPerCurve);
-            for (let jj = 0; jj < nFloatsPerCurve; ++jj) {
-                agentData.push(this.curveData[ii * nFloatsPerCurve + jj]);
-            }
-        }
-        const frameData: VisDataMessage = {
-            msgType: NetMessageEnum.ID_VIS_DATA_ARRIVE,
-            bundleStart: this.currentFrame,
-            bundleSize: 1, // frames
-            bundleData: [
-                {
-                    data: agentData,
-                    frameNumber: this.currentFrame,
-                    time: this.currentFrame,
-                },
-            ],
-            fileName: "hello world",
-        };
-        this.currentFrame++;
-        return frameData;
-    }
-
-    public getInfo(): TrajectoryFileInfoV2 {
-        const typeMapping: EncodedTypeMapping = {};
-        for (let i = 0; i < this.nTypes; ++i) {
-            typeMapping[i] = { name: `fiber${i}` };
-        }
-        return {
-            connId: "hello world",
-            msgType: NetMessageEnum.ID_TRAJECTORY_FILE_INFO,
-            version: 2,
-            timeStepSize: 1,
-            totalSteps: 1000,
-            // bounding volume dimensions
-            size: {
-                x: 12,
-                y: 12,
-                z: 12,
-            },
-            typeMapping: typeMapping,
-            spatialUnits: {
-                magnitude: 1,
-                name: "m",
-            },
-            timeUnits: {
-                magnitude: 1,
-                name: "s",
-            },
-        };
-    }
-}
-
-const curveSim = new CurveSim(1000, 4);
-let curveSimIntervalId = 0;
-
-// these have been set to correspond to backend values
-export const enum NetMessageEnum {
-    ID_UNDEFINED_WEB_REQUEST = 0,
-    ID_VIS_DATA_ARRIVE = 1,
-    ID_VIS_DATA_REQUEST = 2,
-    ID_VIS_DATA_FINISH = 3,
-    ID_VIS_DATA_PAUSE = 4,
-    ID_VIS_DATA_RESUME = 5,
-    ID_VIS_DATA_ABORT = 6,
-    ID_UPDATE_TIME_STEP = 7,
-    ID_UPDATE_RATE_PARAM = 8,
-    ID_MODEL_DEFINITION = 9,
-    ID_HEARTBEAT_PING = 10,
-    ID_HEARTBEAT_PONG = 11,
-    ID_TRAJECTORY_FILE_INFO = 12,
-    ID_GOTO_SIMULATION_TIME = 13,
-    ID_INIT_TRAJECTORY_FILE = 14,
-    // insert new values here before LENGTH
-    LENGTH,
-}
-// these have been set to correspond to backend values
-const enum PlayBackType {
-    ID_LIVE_SIMULATION = 0,
-    ID_PRE_RUN_SIMULATION = 1,
-    ID_TRAJECTORY_FILE_PLAYBACK = 2,
-    // insert new values here before LENGTH
-    LENGTH,
-}
+// setInterval is the playback engine for now
+let simulatorIntervalId = 0;
 
 export class SimulatorConnection extends NetConnection {
     protected logger: ILogger;
@@ -265,51 +75,53 @@ export class SimulatorConnection extends NetConnection {
         //this.logWebSocketRequest(requestDescription, jsonData);
 
         switch (jsonData.msgType) {
-            case NetMessageEnum.ID_UPDATE_TIME_STEP:
+            case ClientMessageEnum.ID_UPDATE_TIME_STEP:
                 break;
-            case NetMessageEnum.ID_UPDATE_RATE_PARAM:
+            case ClientMessageEnum.ID_UPDATE_RATE_PARAM:
                 break;
-            case NetMessageEnum.ID_MODEL_DEFINITION:
+            case ClientMessageEnum.ID_MODEL_DEFINITION:
                 break;
-            case NetMessageEnum.ID_VIS_DATA_REQUEST:
+            case ClientMessageEnum.ID_VIS_DATA_REQUEST:
                 {
                     if (jsonData["frameNumber"]) {
-                        const frame = curveSim.update(jsonData["frameNumber"]);
+                        const frame = localSimulator.update(
+                            jsonData["frameNumber"]
+                        );
                         this.onTrajectoryDataArrive(frame);
                     } else {
-                        const a: TrajectoryFileInfoV2 = curveSim.getInfo();
+                        const a: TrajectoryFileInfoV2 = localSimulator.getInfo();
                         this.onTrajectoryFileInfoArrive(a);
                     }
                 }
                 break;
-            case NetMessageEnum.ID_VIS_DATA_PAUSE:
+            case ClientMessageEnum.ID_VIS_DATA_PAUSE:
                 {
-                    window.clearInterval(curveSimIntervalId);
-                    curveSimIntervalId = 0;
+                    window.clearInterval(simulatorIntervalId);
+                    simulatorIntervalId = 0;
                 }
                 break;
-            case NetMessageEnum.ID_VIS_DATA_RESUME:
+            case ClientMessageEnum.ID_VIS_DATA_RESUME:
                 {
-                    curveSimIntervalId = window.setInterval(() => {
-                        const frame = curveSim.update(0);
+                    simulatorIntervalId = window.setInterval(() => {
+                        const frame = localSimulator.update(0);
                         this.onTrajectoryDataArrive(frame);
                     }, 1);
                 }
                 break;
-            case NetMessageEnum.ID_VIS_DATA_ABORT:
+            case ClientMessageEnum.ID_VIS_DATA_ABORT:
                 {
-                    window.clearInterval(curveSimIntervalId);
-                    curveSimIntervalId = 0;
+                    window.clearInterval(simulatorIntervalId);
+                    simulatorIntervalId = 0;
                 }
                 break;
-            case NetMessageEnum.ID_GOTO_SIMULATION_TIME:
+            case ClientMessageEnum.ID_GOTO_SIMULATION_TIME:
                 {
-                    const frame = curveSim.update(jsonData["time"]);
+                    const frame = localSimulator.update(jsonData["time"]);
                     this.onTrajectoryDataArrive(frame);
                 }
                 break;
-            case NetMessageEnum.ID_INIT_TRAJECTORY_FILE:
-                const a: TrajectoryFileInfoV2 = curveSim.getInfo();
+            case ClientMessageEnum.ID_INIT_TRAJECTORY_FILE:
+                const a: TrajectoryFileInfoV2 = localSimulator.getInfo();
                 console.log("receive trajectory file info");
                 this.onTrajectoryFileInfoArrive(a);
                 break;
@@ -327,7 +139,7 @@ export class SimulatorConnection extends NetConnection {
         }
 
         const jsonData = {
-            msgType: NetMessageEnum.ID_UPDATE_TIME_STEP,
+            msgType: ClientMessageEnum.ID_UPDATE_TIME_STEP,
             timeStep: newTimeStep,
         };
         this.sendSimulationRequest(jsonData, "Update Time-Step");
@@ -339,7 +151,7 @@ export class SimulatorConnection extends NetConnection {
         }
 
         const jsonData = {
-            msgType: NetMessageEnum.ID_UPDATE_RATE_PARAM,
+            msgType: ClientMessageEnum.ID_UPDATE_RATE_PARAM,
             paramName: paramName,
             paramValue: paramValue,
         };
@@ -353,7 +165,7 @@ export class SimulatorConnection extends NetConnection {
 
         const dataToSend = {
             model: model,
-            msgType: NetMessageEnum.ID_MODEL_DEFINITION,
+            msgType: ClientMessageEnum.ID_MODEL_DEFINITION,
         };
         this.sendSimulationRequest(dataToSend, "Model Definition");
     }
@@ -369,8 +181,8 @@ export class SimulatorConnection extends NetConnection {
      */
     public startRemoteSimPreRun(timeStep: number, numTimeSteps: number): void {
         const jsonData = {
-            msgType: NetMessageEnum.ID_VIS_DATA_REQUEST,
-            mode: PlayBackType.ID_PRE_RUN_SIMULATION,
+            msgType: ClientMessageEnum.ID_VIS_DATA_REQUEST,
+            mode: ClientPlayBackType.ID_PRE_RUN_SIMULATION,
             timeStep: timeStep,
             numTimeSteps: numTimeSteps,
         };
@@ -382,8 +194,8 @@ export class SimulatorConnection extends NetConnection {
 
     public startRemoteSimLive(): void {
         const jsonData = {
-            msgType: NetMessageEnum.ID_VIS_DATA_REQUEST,
-            mode: PlayBackType.ID_LIVE_SIMULATION,
+            msgType: ClientMessageEnum.ID_VIS_DATA_REQUEST,
+            mode: ClientPlayBackType.ID_LIVE_SIMULATION,
         };
 
         this.connectToRemoteServer(this.getIp()).then(() => {
@@ -393,8 +205,8 @@ export class SimulatorConnection extends NetConnection {
 
     public startRemoteTrajectoryPlayback(fileName: string): Promise<void> {
         const jsonData = {
-            msgType: NetMessageEnum.ID_VIS_DATA_REQUEST,
-            mode: PlayBackType.ID_TRAJECTORY_FILE_PLAYBACK,
+            msgType: ClientMessageEnum.ID_VIS_DATA_REQUEST,
+            mode: ClientPlayBackType.ID_TRAJECTORY_FILE_PLAYBACK,
             "file-name": fileName,
         };
 
@@ -411,7 +223,7 @@ export class SimulatorConnection extends NetConnection {
             return;
         }
         this.sendSimulationRequest(
-            { msgType: NetMessageEnum.ID_VIS_DATA_PAUSE },
+            { msgType: ClientMessageEnum.ID_VIS_DATA_PAUSE },
             "Pause Simulation"
         );
     }
@@ -421,7 +233,7 @@ export class SimulatorConnection extends NetConnection {
             return;
         }
         this.sendSimulationRequest(
-            { msgType: NetMessageEnum.ID_VIS_DATA_RESUME },
+            { msgType: ClientMessageEnum.ID_VIS_DATA_RESUME },
             "Resume Simulation"
         );
     }
@@ -431,7 +243,7 @@ export class SimulatorConnection extends NetConnection {
             return;
         }
         this.sendSimulationRequest(
-            { msgType: NetMessageEnum.ID_VIS_DATA_ABORT },
+            { msgType: ClientMessageEnum.ID_VIS_DATA_ABORT },
             "Abort Simulation"
         );
     }
@@ -439,8 +251,8 @@ export class SimulatorConnection extends NetConnection {
     public requestSingleFrame(startFrameNumber: number): void {
         this.sendSimulationRequest(
             {
-                msgType: NetMessageEnum.ID_VIS_DATA_REQUEST,
-                mode: PlayBackType.ID_TRAJECTORY_FILE_PLAYBACK,
+                msgType: ClientMessageEnum.ID_VIS_DATA_REQUEST,
+                mode: ClientPlayBackType.ID_TRAJECTORY_FILE_PLAYBACK,
                 frameNumber: startFrameNumber,
             },
             "Request Single Frame"
@@ -450,7 +262,7 @@ export class SimulatorConnection extends NetConnection {
     public gotoRemoteSimulationTime(timeNanoSeconds: number): void {
         this.sendSimulationRequest(
             {
-                msgType: NetMessageEnum.ID_GOTO_SIMULATION_TIME,
+                msgType: ClientMessageEnum.ID_GOTO_SIMULATION_TIME,
                 time: timeNanoSeconds,
             },
             "Load single frame at specified Time"
@@ -460,7 +272,7 @@ export class SimulatorConnection extends NetConnection {
     public requestTrajectoryFileInfo(fileName: string): void {
         this.sendSimulationRequest(
             {
-                msgType: NetMessageEnum.ID_INIT_TRAJECTORY_FILE,
+                msgType: ClientMessageEnum.ID_INIT_TRAJECTORY_FILE,
                 fileName: fileName,
             },
             "Initialize trajectory file info"
